Add styles to nav.css
