+++ src/Nav.css
@@ -1,26 +1,35 @@
 .title {
- font-size: 36px;
+  font-size: 36px;
+}
+
+h1 {
+  text-align: center;
 }
 
 nav {
- padding: 2rem;
- display: flex;
- justify-content: space-between;
- align-items: baseline;
+  background-color: darkgoldenrod;
+  padding: 2rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: baseline;
 }
 
 nav ul {
- width: 260px;
- display: flex;
- justify-content: space-between;
+  width: 260px;
+  display: flex;
+  justify-content: space-between;
+  font-style: italic;
+  color: white;
+  list-style-type: none;
+  text-decoration: none;
 }
 
 nav li {
- font-size: 22px;
+  font-size: 22px;
+  background-color: darkred;
+  list-style: none;
 }
 
 .middle {
- border-left: 2px solid blue;
- border-right: 2px solid blue;
- padding: 0 0.8rem;
+  padding: 0 0.8rem;
 }
