+++ src/components/Calculator.css
@@ -1,10 +1,3 @@
-.calculator {
-  max-width: 400px;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 10rem;
-}
-
 .calculator-result {
   background-color: grey;
   text-align: end;
