+++ src/index.css
@@ -14,6 +14,10 @@ ul {
   list-style: none;
 }
 
+body {
+  background-color: teal;
+}
+
 /*
     ==============================================
                 ****** Global *****
