+++ src/pages/Home.css
@@ -1,0 +1,14 @@
+.home-holder {
+  padding: 0 5rem;
+  margin-top: 3rem;
+}
+
+.home-title {
+  font-size: 26px;
+}
+
+.home-text {
+  margin-left: 3rem;
+  margin-top: 5rem;
+  font-size: 20px;
+}
