+++ src/pages/Home.js
@@ -1,0 +1,24 @@
+import React from 'react';
+import './Home.css';
+
+export default function Home() {
+  return (
+    <div className="home-holder">
+      <h2 className="home-title">Welcome to my page!</h2>
+      <p className="home-text">
+        Lorem ipsum dolor sit amet consectetur adipisicing elit. Aperiam quis
+        deserunt ut. Expedita delectus, ipsam aperiam iure optio hic praesentium
+        modi repellat in nam ut soluta dolore ex illum nisi. Lorem ipsum dolor
+        sit amet consectetur adipisicing elit. Aperiam quis deserunt ut.
+        Expedita delectus, ipsam aperiam iure optio hic praesentium modi
+        repellat in nam ut soluta dolore ex illum nisi. Lorem ipsum dolor sit
+        amet consectetur adipisicing elit. Aperiam quis deserunt ut. Expedita
+        delectus, ipsam aperiam iure optio hic praesentium modi repellat in nam
+        ut soluta dolore ex illum nisi. Lorem ipsum dolor sit amet consectetur
+        adipisicing elit. Aperiam quis deserunt ut. Expedita delectus, ipsam
+        aperiam iure optio hic praesentium modi repellat in nam ut soluta dolore
+        ex illum nisi.
+      </p>
+    </div>
+  );
+}
