+++ src/pages/Quote.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import './Quote.css';
+
+export default function Quote() {
+  return (
+    <p className="quote">
+      Do not worry too much about your difficulties in mathematics
+      I can assure you that mine are still greater.
+      Pure mathematics is, in its way, the poetry of logical ideas.
+    </p>
+  );
+}
