--- src/single page app/Quote.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import './Quote.css';
-
-export default function Quote() {
-  return (
-    <p className="quote">
-      The essence of math is not to make simple things complicated, but to make
-      complicated things simple. — Stan Gudder
-    </p>
-  );
-}
